services/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv